+++ src/lib/helper.ts
@@ -39,3 +39,39 @@ export function getFromSessionStorage(key: string): string | null {
   }
   return null;
 }
+
+export function classNames(...classes: any[]) {
+  return classes.filter(Boolean).join(' ');
+}
+
+export const getExampleRequest = (pkg: string, method: Method): string => {
+  return JSON.stringify(
+    {
+      id: 1,
+      jsonrpc: '2.0',
+      method: pkg + '.' + method.name,
+      params: method.params.map((param) =>
+        param.schema && param.schema.examples
+          ? param.schema.examples[0]
+          : undefined
+      ),
+    },
+    null,
+    2
+  );
+};
+
+export const getExampleResponse = (method: Method): string => {
+  return JSON.stringify(
+    {
+      id: 1,
+      jsonrpc: '2.0',
+      result:
+        method.result.description == 'Null' || !method.result.schema.examples
+          ? []
+          : [method.result.schema.examples[0]],
+    },
+    null,
+    2
+  );
+};
