+++ src/lib/types.ts
@@ -1,0 +1,26 @@
+export interface INotification {
+  message: string;
+  success: boolean;
+  active: boolean;
+}
+
+export interface NodeError {
+  code: number;
+  message: string;
+}
+
+export type Param = {
+  name: string;
+  description: string;
+  schema: any;
+};
+
+export type Method = {
+  name: string;
+  description: string;
+  auth: string;
+  params: Param[];
+  result: Param;
+};
+
+export type MethodByPkg = { [key: string]: Method[] };
