+++ src/pages/index.tsx
@@ -4,30 +4,24 @@
 /* eslint-disable @next/next/no-img-element */
 import {
   Dialog,
-  Transition,
-  TransitionChild,
   DialogPanel,
   DialogTitle,
+  Transition,
+  TransitionChild,
 } from '@headlessui/react';
-import {
-  Bars3BottomLeftIcon,
-  XMarkIcon,
-  CheckCircleIcon,
-  XCircleIcon,
-} from '@heroicons/react/24/outline';
-import {
-  ChevronDownIcon,
-  ChevronRightIcon,
-  CommandLineIcon,
-} from '@heroicons/react/24/solid';
+import { Bars3BottomLeftIcon, XMarkIcon } from '@heroicons/react/24/outline';
 import axios from 'axios';
-import { AxiosError } from 'axios';
 import { Fragment, useEffect, useState } from 'react';
-import { useRef } from 'react';
 import { CopyToClipboard } from 'react-copy-to-clipboard';
 import { FiClipboard } from 'react-icons/fi';
 import SyntaxHighlighter from 'react-syntax-highlighter';
-import { Editor } from '@monaco-editor/react';
+
+import { classNames } from '@/lib/helper';
+import { INotification, MethodByPkg, Param } from '@/lib/types';
+
+import NotificationModal from '@/components/NotificationModal';
+import Playground from '@/components/Playground';
+import RPCMethod from '@/components/RPCMethod';
 import Seo from '@/components/Seo';
 
 const clients = [
@@ -52,32 +46,6 @@ const clients = [
     current: false,
   },
 ];
-
-const tabs = [
-  { name: 'Request', href: '#' },
-  { name: 'Response', href: '#' },
-  { name: 'Configuration', href: '#' },
-];
-
-function classNames(...classes: any[]) {
-  return classes.filter(Boolean).join(' ');
-}
-
-type Param = {
-  name: string;
-  description: string;
-  schema: any;
-};
-
-type Method = {
-  name: string;
-  description: string;
-  auth: string;
-  params: Param[];
-  result: Param;
-};
-
-type MethodByPkg = { [key: string]: Method[] };
 
 function extractAuth(methodDescription: string): string {
   return methodDescription.split('Auth level: ')[1];
@@ -133,17 +101,6 @@ const versions = [
   `v0.13.5`,
 ].reverse();
 
-interface INotification {
-  message: string;
-  success: boolean;
-  active: boolean;
-}
-
-interface NodeError {
-  code: number;
-  message: string;
-}
-
 export default function Example() {
   const handleVersionChange = (event: React.ChangeEvent<HTMLSelectElement>) => {
     setSelectedVersion(event.target.value);
@@ -159,7 +116,6 @@ export default function Example() {
   });
 
   const [currentRequest, setCurrentRequest] = useState<string>('');
-  const [currentResponse, setCurrentResponse] = useState<string>('');
   const [playgroundOpen, setPlaygroundOpen] = useState(false);
   const [notification, setNotification] = useState<INotification>({
     message: '',
@@ -167,14 +123,9 @@ export default function Example() {
     active: false,
   });
 
-  // request: 0, response: 1, config: 2
-  const [currentTab, setCurrentTab] = useState(0);
   const [selectedVersion, setSelectedVersion] = useState(versions[0]);
   const [showHash, setShowHash] = useState('');
   const [searchTerm, setSearchTerm] = useState('');
-
-  const [hostname, setHostname] = useState('');
-  const [authToken, setAuthToken] = useState('');
 
   useEffect(() => {
     const fetchJsonData = async (version: string) => {
@@ -234,13 +185,14 @@ export default function Example() {
     <>
       <Seo />
       <div>
-        <Transition.Root show={sidebarOpen} as={Fragment}>
+        {/* SIDEBAR MOBILE */}
+        <Transition show={sidebarOpen} as={Fragment}>
           <Dialog
             as='div'
             className='relative z-40 md:hidden'
             onClose={setSidebarOpen}
           >
-            <Transition.Child
+            <TransitionChild
               as={Fragment}
               enter='transition-opacity ease-linear duration-300'
               enterFrom='opacity-0'
@@ -250,10 +202,10 @@ export default function Example() {
               leaveTo='opacity-0'
             >
               <div className='fixed inset-0 bg-gray-600 bg-opacity-75' />
-            </Transition.Child>
+            </TransitionChild>
 
             <div className='fixed inset-0 z-40 flex'>
-              <Transition.Child
+              <TransitionChild
                 as={Fragment}
                 enter='transition ease-in-out duration-300 transform'
                 enterFrom='-translate-x-full'
@@ -262,8 +214,8 @@ export default function Example() {
                 leaveFrom='translate-x-0'
                 leaveTo='-translate-x-full'
               >
-                <Dialog.Panel className='relative flex w-full max-w-xs flex-1 flex-col bg-white pb-4 pt-5'>
-                  <Transition.Child
+                <DialogPanel className='relative flex w-full max-w-xs flex-1 flex-col bg-white pb-4 pt-5'>
+                  <TransitionChild
                     as={Fragment}
                     enter='ease-in-out duration-300'
                     enterFrom='opacity-0'
@@ -285,7 +237,7 @@ export default function Example() {
                         />
                       </button>
                     </div>
-                  </Transition.Child>
+                  </TransitionChild>
                   <div className='logo-container border-b border-gray-200 pb-4 md:border-none md:pb-0'>
                     <div className='flex flex-shrink-0 items-center px-4'>
                       <img
@@ -338,16 +290,16 @@ export default function Example() {
                       ))}
                     </nav>
                   </div>
-                </Dialog.Panel>
-              </Transition.Child>
+                </DialogPanel>
+              </TransitionChild>
               <div className='w-14 flex-shrink-0'>
                 {/* Dummy element to force sidebar to shrink to fit close icon */}
               </div>
             </div>
           </Dialog>
-        </Transition.Root>
+        </Transition>
 
-        {/* Static sidebar for desktop */}
+        {/* SIDEBAR DESKTOP */}
         <div className='relative hidden md:fixed md:inset-y-0 md:flex md:w-64 md:flex-col'>
           {/* Sidebar component, swap this element with another sidebar if you like */}
           <div className='flex flex-grow flex-col overflow-y-auto border-r border-gray-200 bg-white pt-5'>
@@ -458,6 +410,7 @@ export default function Example() {
 
         <div className='md:pl-64'>
           <div className='mx-auto flex max-w-4xl flex-col md:px-8 xl:px-0'>
+            {/* TOP NAVBAR */}
             <div className='sticky top-0 z-10 flex h-16 flex-shrink-0 border-b border-gray-200 bg-white md:hidden'>
               <button
                 type='button'
@@ -605,20 +558,9 @@ export default function Example() {
                                     method={method}
                                     activateSidebar={activateSidebar}
                                     selectedVersion={selectedVersion}
+                                    setCurrentRequest={setCurrentRequest}
+                                    setPlaygroundOpen={setPlaygroundOpen}
                                   />
-                                  <button
-                                    type='button'
-                                    className='mt-3 inline-flex w-full justify-center rounded-md border border-gray-300 bg-white px-4 py-2 text-base font-medium text-gray-700 shadow-sm hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-purple-500 focus:ring-offset-2 sm:mt-0 sm:w-auto sm:text-sm'
-                                    onClick={() => {
-                                      pkg &&
-                                        setCurrentRequest(
-                                          getExampleRequest(pkg, method)
-                                        );
-                                      setPlaygroundOpen(true);
-                                    }}
-                                  >
-                                    Try it out
-                                  </button>
                                 </div>
                               ))}
                             </div>
@@ -631,606 +573,116 @@ export default function Example() {
           </div>
         </div>
       </div>
-      {/* EXAMPLE TYPE MODAL */}
-      <Transition show={open}>
-        <Dialog as='div' className='relative z-10' onClose={setOpen}>
-          <TransitionChild
-            as={Fragment}
-            enter='ease-out duration-300'
-            enterFrom='opacity-0'
-            enterTo='opacity-100'
-            leave='ease-in duration-200'
-            leaveFrom='opacity-100'
-            leaveTo='opacity-0'
-          >
-            <div className='fixed inset-0 bg-gray-500 bg-opacity-75 transition-opacity' />
-          </TransitionChild>
 
-          <div className='fixed inset-0 z-10 overflow-y-auto'>
-            <div className='flex min-h-full items-end justify-center p-4 text-center sm:items-center sm:p-0'>
-              <TransitionChild
-                enter='ease-out duration-300'
-                enterFrom='opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95'
-                enterTo='opacity-100 translate-y-0 sm:scale-100'
-                leave='ease-in duration-200'
-                leaveFrom='opacity-100 translate-y-0 sm:scale-100'
-                leaveTo='opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95'
-              >
-                <DialogPanel className='relative transform rounded-lg bg-white px-4 pb-4 pt-5 text-left shadow-xl transition-all sm:my-8 sm:w-full sm:max-w-lg sm:p-6'>
-                  <div className='absolute right-0 top-0 hidden pr-4 pt-4 sm:block'>
-                    <button
-                      type='button'
-                      className='rounded-md bg-white text-gray-400 hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-purple-500 focus:ring-offset-2'
-                      onClick={() => setOpen(false)}
-                    >
-                      <span className='sr-only'>Close</span>
-                      <XMarkIcon className='h-6 w-6' aria-hidden='true' />
-                    </button>
-                  </div>
-                  <div className='overflow-x-auto sm:flex sm:items-start'>
-                    <div className='mt-3 text-center sm:ml-4 sm:mt-0 sm:text-left'>
-                      <DialogTitle
-                        as='h3'
-                        className='text-lg font-medium leading-6 text-gray-900'
-                      >
-                        {currentParam.description}
-                      </DialogTitle>
-                      <div className='mt-2'>
-                        {currentParam.schema?.examples &&
-                          currentParam.schema.examples.length > 0 && (
-                            <p className='text-sm text-gray-500'>
-                              Example Value:
-                              <SyntaxHighlighter
-                                language='javascript'
-                                customStyle={{
-                                  backgroundColor: 'transparent',
-                                }}
-                              >
-                                {JSON.stringify(
-                                  currentParam.schema.examples[0],
-                                  null,
-                                  '\t'
-                                )}
-                              </SyntaxHighlighter>
-                            </p>
-                          )}
-                      </div>
-                    </div>
-                  </div>
-                  <div className='mt-5 sm:mt-4 sm:flex sm:flex-row-reverse'>
-                    <button
-                      type='button'
-                      className='mt-3 inline-flex w-full justify-center rounded-md border border-gray-300 bg-white px-4 py-2 text-base font-medium text-gray-700 shadow-sm hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-purple-500 focus:ring-offset-2 sm:mt-0 sm:w-auto sm:text-sm'
-                      onClick={() => setOpen(false)}
-                    >
-                      Close
-                    </button>
-                  </div>
-                </DialogPanel>
-              </TransitionChild>
-            </div>
-          </div>
-        </Dialog>
-      </Transition>
-      <div className='relative'>
-        {/* PLAYGROUND MODAL */}
-        <Transition show={playgroundOpen}>
-          <Dialog className='relative' onClose={console.log}>
-            <TransitionChild
-              enter='ease-out duration-300'
-              enterFrom='opacity-0'
-              enterTo='opacity-100'
-              leave='ease-in duration-200'
-              leaveFrom='opacity-100'
-              leaveTo='opacity-0'
-            >
-              <div className='pointer-events-none fixed inset-0 bg-gray-500 bg-opacity-75 transition-opacity' />
-            </TransitionChild>
-
-            <div className='pointer-events-none fixed inset-0 w-screen overflow-y-auto'>
-              <div className='pointer-events-auto flex min-h-full items-end justify-center p-4 text-center sm:items-center sm:p-0'>
-                <TransitionChild
-                  enter='ease-out duration-300'
-                  enterFrom='opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95'
-                  enterTo='opacity-100 translate-y-0 sm:scale-100'
-                  leave='ease-in duration-200'
-                  leaveFrom='opacity-100 translate-y-0 sm:scale-100'
-                  leaveTo='opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95'
-                >
-                  <DialogPanel className='relative transform overflow-hidden rounded-lg bg-white px-4 pb-4 pt-5 text-left shadow-xl transition-all sm:my-8 sm:w-full sm:max-w-2xl sm:p-6'>
-                    <div className='sm:flex-grow'>
-                      <div className='flex'>
-                        <div className='mx-auto flex h-12 w-12 flex-shrink-0 items-center justify-center rounded-full bg-purple-100 sm:mx-0 sm:h-10 sm:w-10'>
-                          <CommandLineIcon
-                            className='h-6 w-6 text-purple-600'
-                            aria-hidden='true'
-                          />
-                        </div>
-                        <DialogTitle
-                          as='h3'
-                          className='ml-3 mt-2 text-base font-semibold text-gray-900'
-                        >
-                          Node Playground
-                        </DialogTitle>
-                      </div>
-                      <div className='mt-3 text-center sm:ml-4 sm:mt-0 sm:text-left'>
-                        <div className='mt-2 flex-grow flex-row'>
-                          {/* TABS */}
-                          <div className='sm:hidden'>
-                            <label htmlFor='tabs' className='sr-only'>
-                              Select a tab
-                            </label>
-                            {/* Use an "onChange" listener to redirect the user to the selected tab URL. */}
-                            <select
-                              id='tabs'
-                              name='tabs'
-                              className='block w-full rounded-md border-gray-300 focus:border-purple-500 focus:ring-purple-500'
-                              value={tabs[currentTab].name}
-                              onChange={(e) => {
-                                switch (e.currentTarget.value) {
-                                  case 'Request':
-                                    setCurrentTab(0);
-                                    break;
-                                  case 'Response':
-                                    setCurrentTab(1);
-                                    break;
-                                  case 'Configure':
-                                    setCurrentTab(2);
-                                    break;
-                                  default:
-                                    setCurrentTab(0);
-                                    break;
-                                }
-                              }}
-                            >
-                              {tabs.map((tab) => (
-                                <option key={tab.name}>{tab.name}</option>
-                              ))}
-                            </select>
-                          </div>
-                          <div className='hidden flex-grow sm:block'>
-                            <nav
-                              className='isolate flex divide-x divide-gray-200 rounded-lg shadow'
-                              aria-label='Tabs'
-                            >
-                              {tabs.map((tab, tabIdx) => (
-                                <a
-                                  key={tab.name}
-                                  href={tab.href}
-                                  onClick={(_) => setCurrentTab(tabIdx)}
-                                  className={classNames(
-                                    currentTab == tabIdx
-                                      ? 'text-gray-900'
-                                      : 'text-gray-500 hover:text-gray-700',
-                                    tabIdx === 0 ? 'rounded-l-lg' : '',
-                                    tabIdx === tabs.length - 1
-                                      ? 'rounded-r-lg'
-                                      : '',
-                                    'group relative min-w-0 flex-1 overflow-hidden bg-white px-4 py-4 text-center text-sm font-medium hover:bg-gray-50 focus:z-10'
-                                  )}
-                                  aria-current={
-                                    currentTab == tabIdx ? 'page' : undefined
-                                  }
-                                >
-                                  <span>{tab.name}</span>
-                                  <span
-                                    aria-hidden='true'
-                                    className={classNames(
-                                      currentTab == tabIdx
-                                        ? 'bg-purple-500'
-                                        : 'bg-transparent',
-                                      'absolute inset-x-0 bottom-0 h-0.5'
-                                    )}
-                                  />
-                                </a>
-                              ))}
-                            </nav>
-                          </div>
-                          {/* PLAYGROUND */}
-                          {currentTab == 0 && (
-                            <Editor
-                              language='json'
-                              options={{
-                                scrollBeyondLastLine: false,
-                                minimap: { enabled: false },
-                                useShadows: false,
-                              }}
-                              className='mt-3 min-h-52'
-                              value={currentRequest}
-                              onChange={(value) =>
-                                value && setCurrentRequest(value)
-                              }
-                            />
-                          )}
-                          {currentTab == 1 && (
-                            <Editor
-                              language='json'
-                              options={{
-                                scrollBeyondLastLine: false,
-                                minimap: { enabled: false },
-                                useShadows: false,
-                                readOnly: true,
-                              }}
-                              className='mt-3 min-h-52'
-                              value={currentResponse}
-                            />
-                          )}
-                          {currentTab == 2 && (
-                            <div>
-                              <label
-                                htmlFor='ip'
-                                className='mt-6 block text-sm font-medium leading-6 text-gray-900'
-                              >
-                                IP Address
-                              </label>
-                              <div className='mt-2 flex rounded-md shadow-sm'>
-                                <span className='inline-flex items-center rounded-l-md border border-r-0 border-gray-300 px-3 text-gray-500 sm:text-sm'>
-                                  http://
-                                </span>
-                                <input
-                                  type='text'
-                                  name='ip'
-                                  id='ip'
-                                  className='block w-full min-w-0 flex-1 rounded-none rounded-r-md border-0 py-1.5 text-gray-900 ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6'
-                                  placeholder='localhost:26658'
-                                  value={hostname}
-                                  onChange={(e) => setHostname(e.target.value)}
-                                />
-                              </div>
-                              <p
-                                className='mt-2 text-sm text-gray-500'
-                                id='protocol-description'
-                              >
-                                ws:// or wss:// is not supported at this time
-                              </p>
-                              <label
-                                htmlFor='authtoken'
-                                className='mt-6 block text-sm font-medium leading-6 text-gray-900'
-                              >
-                                Auth Token
-                              </label>
-                              <div className='mt-2 flex rounded-md shadow-sm'>
-                                <input
-                                  type='text'
-                                  name='authtoken'
-                                  id='authtoken'
-                                  className='block w-full min-w-0 flex-1 rounded-md border-0 py-1.5 text-gray-900 ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6'
-                                  placeholder=''
-                                  value={authToken}
-                                  onChange={(e) => setAuthToken(e.target.value)}
-                                />
-                              </div>
-                              <p
-                                className='mt-2 text-sm text-gray-500'
-                                id='auth-description'
-                              >
-                                Only set this if you don't have the
-                                --rpc.skip-auth flag set.
-                              </p>
-                            </div>
-                          )}
-                        </div>
-                      </div>
-                    </div>
-                    <div className='mt-5 sm:mt-4 sm:flex sm:flex-row-reverse'>
-                      <button
-                        type='button'
-                        className='mt-3 inline-flex w-full justify-center rounded-md bg-purple-100 px-3 py-2 text-sm font-semibold text-purple-900 shadow-sm ring-1 ring-inset ring-purple-300 hover:bg-gray-50 sm:mt-0 sm:w-auto'
-                        onClick={async () => {
-                          try {
-                            const data = await sendRequest(
-                              currentRequest,
-                              hostname,
-                              authToken
-                            );
-                            setCurrentResponse(JSON.stringify(data, null, 2));
-                            setCurrentTab(1);
-                            if ('error' in data) {
-                              setNotification({
-                                active: true,
-                                success: false,
-                                message: (data.error as NodeError).message,
-                              });
-                            } else {
-                              setNotification({
-                                active: true,
-                                success: true,
-                                message: 'Request sent successfully',
-                              });
-                            }
-                          } catch (e) {
-                            setNotification({
-                              active: true,
-                              success: false,
-                              message: (e as AxiosError).message,
-                            });
-                            return;
-                          }
-                        }}
-                        data-autofocus
-                      >
-                        Send Request
-                      </button>
-                      <button
-                        type='button'
-                        className='mr-3 mt-3 inline-flex w-full justify-center rounded-md bg-white px-3 py-2 text-sm font-semibold text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 hover:bg-gray-50 sm:mt-0 sm:w-auto'
-                        onClick={() => setPlaygroundOpen(false)}
-                        data-autofocus
-                      >
-                        Dismiss
-                      </button>
-                    </div>
-                  </DialogPanel>
-                </TransitionChild>
-              </div>
-            </div>
-          </Dialog>
-        </Transition>
-        {/* NOTIFICATIONS */}
-        {/* I use playground open because the css is janky otherwise (due to pointer events and z index) */}
-        {playgroundOpen && (
-          <div
-            aria-live='assertive'
-            className='pointer-events-auto fixed inset-0 z-50 flex items-end px-4 py-6 sm:items-start sm:p-6'
-          >
-            <div className='z-50 flex w-full flex-col items-center space-y-4 sm:items-end'>
-              <Transition
-                show={notification.active}
-                enter='transform ease-out duration-300 transition'
-                enterFrom='translate-y-2 opacity-0 sm:translate-y-0 sm:translate-x-2'
-                enterTo='translate-y-0 opacity-100 sm:translate-x-0 z-50'
-                leave='transition ease-in duration-100'
-                leaveFrom='opacity-100'
-                leaveTo='opacity-0'
-              >
-                <div className='z-50 w-full max-w-sm overflow-hidden rounded-lg bg-white shadow-lg ring-1 ring-black ring-opacity-5'>
-                  <div className='z-50 p-4'>
-                    <div className='z-50 flex items-start'>
-                      <div className='z-50 flex-shrink-0'>
-                        {notification.success ? (
-                          <CheckCircleIcon
-                            className='h-6 w-6 text-green-400'
-                            aria-hidden='true'
-                          />
-                        ) : (
-                          <XCircleIcon
-                            className='h-6 w-6 text-red-400'
-                            aria-hidden='true'
-                          />
-                        )}
-                      </div>
-                      <div className='ml-3 w-0 flex-1 pt-0.5'>
-                        <p className='text-sm font-medium text-gray-900'>
-                          {notification.success
-                            ? 'Request Succeeded'
-                            : 'Request Failed With Error:'}
-                        </p>
-                        <p className='mt-1 text-sm text-gray-500'>
-                          {notification.message}
-                        </p>
-                      </div>
-                      <div className='z-50 ml-4 flex flex-shrink-0'>
-                        <button
-                          type='button'
-                          className='z-50 inline-flex rounded-md bg-white text-gray-400 hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2'
-                          onClick={() => {
-                            setNotification({ ...notification, active: false });
-                          }}
-                        >
-                          <span className='sr-only'>Close</span>
-                          <XMarkIcon
-                            className='z-50 h-5 w-5'
-                            aria-hidden='true'
-                          />
-                        </button>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </Transition>
-            </div>
-          </div>
-        )}
-      </div>
+      <ExampleTypeModal
+        open={open}
+        setOpen={setOpen}
+        currentParam={currentParam}
+      />
+      <Playground
+        playgroundOpen={playgroundOpen}
+        setPlaygroundOpen={setPlaygroundOpen}
+        currentRequest={currentRequest}
+        setCurrentRequest={setCurrentRequest}
+        setNotification={setNotification}
+      />
+      <NotificationModal
+        notification={notification}
+        setNotification={setNotification}
+      />
     </>
   );
 }
 
-const RPCMethod = ({
-  pkg,
-  method,
-  activateSidebar,
-  selectedVersion,
+const ExampleTypeModal = ({
+  open,
+  setOpen,
+  currentParam,
 }: {
-  pkg: string;
-  method: Method;
-  activateSidebar: (param: Param) => void;
-  selectedVersion: string;
+  open: boolean;
+  setOpen: (open: boolean) => void;
+  currentParam: Param;
 }) => {
-  const [showRequest, setShowRequest] = useState(false);
-  const [showResponse, setShowResponse] = useState(false);
-  const methodRef = useRef<HTMLDivElement>(null); // create a ref
-
-  const handleCopyClick = () => {
-    const hash = window.location.hash.substring(1);
-    const element = document.getElementById(hash);
-    element?.scrollIntoView();
-  };
-
   return (
-    <div
-      key={`${pkg}.${method.name}`}
-      id={`${pkg}.${method.name}`}
-      className='py-2'
-      ref={methodRef}
-    >
-      <div className='flex items-center justify-between'>
-        <p className='text-md'>
-          {method.name}(
-          {method.params.map((param, i, { length }) => (
-            <span key={param.name} className='text-sm text-gray-700'>
-              <span>{param.name}</span>{' '}
-              <span
-                className='text-blue-500 hover:cursor-pointer hover:font-bold'
-                onClick={() => activateSidebar(param)}
-              >
-                {param.description}
-                {length - 1 != i && ', '}
-              </span>
-            </span>
-          ))}
-          )
-          {method.result.description != 'Null' && (
-            <span
-              className='ml-2 text-sm text-blue-500 hover:cursor-pointer hover:font-bold'
-              onClick={() => activateSidebar(method.result)}
-            >
-              {method.result.description}
-            </span>
-          )}
-          <span className='ml-2 inline-flex items-center rounded-full bg-purple-100 px-2.5 py-0.5 text-xs font-medium text-purple-800'>
-            perms: {method.auth}
-          </span>
-        </p>
-        <CopyToClipboard
-          text={`${window.location.origin}/?version=${selectedVersion}#${pkg}.${method.name}`}
-          onCopy={handleCopyClick}
+    <Transition show={open}>
+      <Dialog as='div' className='relative z-10' onClose={setOpen}>
+        <TransitionChild
+          as={Fragment}
+          enter='ease-out duration-300'
+          enterFrom='opacity-0'
+          enterTo='opacity-100'
+          leave='ease-in duration-200'
+          leaveFrom='opacity-100'
+          leaveTo='opacity-0'
         >
-          <span
-            className='cursor-pointer text-gray-500 hover:text-blue-500'
-            onClick={() => {
-              window.location.hash = `${pkg}.${method.name}`;
-            }}
-          >
-            #
-          </span>
-        </CopyToClipboard>
-      </div>
-      <p className='text-sm font-light text-gray-700'>{method.description}</p>
-      <div className='mt-6 overflow-hidden rounded-lg bg-white text-sm shadow'>
-        <div
-          className='flex px-4 py-5 hover:cursor-pointer hover:bg-gray-50 sm:px-6'
-          onClick={() => setShowRequest(!showRequest)}
-        >
-          {showRequest ? (
-            <ChevronDownIcon className='mr-2 h-5 w-5' aria-hidden='true' />
-          ) : (
-            <ChevronRightIcon className='mr-2 h-5 w-5' aria-hidden='true' />
-          )}
-          Request
-        </div>
-        {showRequest && (
-          <div className='bg-gray-50 px-4 py-5 text-sm sm:p-6'>
-            <SyntaxHighlighter
-              language='javascript'
-              customStyle={{
-                backgroundColor: 'transparent',
-              }}
+          <div className='fixed inset-0 bg-gray-500 bg-opacity-75 transition-opacity' />
+        </TransitionChild>
+
+        <div className='fixed inset-0 z-10 overflow-y-auto'>
+          <div className='flex min-h-full items-end justify-center p-4 text-center sm:items-center sm:p-0'>
+            <TransitionChild
+              enter='ease-out duration-300'
+              enterFrom='opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95'
+              enterTo='opacity-100 translate-y-0 sm:scale-100'
+              leave='ease-in duration-200'
+              leaveFrom='opacity-100 translate-y-0 sm:scale-100'
+              leaveTo='opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95'
             >
-              {JSON.stringify(
-                {
-                  id: 1,
-                  jsonrpc: '2.0',
-                  method: pkg + '.' + method.name,
-                  params: method.params.map((param) =>
-                    param.schema && param.schema.examples
-                      ? param.schema.examples[0]
-                      : undefined
-                  ),
-                },
-                null,
-                2
-              )}
-            </SyntaxHighlighter>
+              <DialogPanel className='relative transform rounded-lg bg-white px-4 pb-4 pt-5 text-left shadow-xl transition-all sm:my-8 sm:w-full sm:max-w-lg sm:p-6'>
+                <div className='absolute right-0 top-0 hidden pr-4 pt-4 sm:block'>
+                  <button
+                    type='button'
+                    className='rounded-md bg-white text-gray-400 hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-purple-500 focus:ring-offset-2'
+                    onClick={() => setOpen(false)}
+                  >
+                    <span className='sr-only'>Close</span>
+                    <XMarkIcon className='h-6 w-6' aria-hidden='true' />
+                  </button>
+                </div>
+                <div className='overflow-x-auto sm:flex sm:items-start'>
+                  <div className='mt-3 text-center sm:ml-4 sm:mt-0 sm:text-left'>
+                    <DialogTitle
+                      as='h3'
+                      className='text-lg font-medium leading-6 text-gray-900'
+                    >
+                      {currentParam.description}
+                    </DialogTitle>
+                    <div className='mt-2'>
+                      {currentParam.schema?.examples &&
+                        currentParam.schema.examples.length > 0 && (
+                          <p className='text-sm text-gray-500'>
+                            Example Value:
+                            <SyntaxHighlighter
+                              language='javascript'
+                              customStyle={{
+                                backgroundColor: 'transparent',
+                              }}
+                            >
+                              {JSON.stringify(
+                                currentParam.schema.examples[0],
+                                null,
+                                '\t'
+                              )}
+                            </SyntaxHighlighter>
+                          </p>
+                        )}
+                    </div>
+                  </div>
+                </div>
+                <div className='mt-5 sm:mt-4 sm:flex sm:flex-row-reverse'>
+                  <button
+                    type='button'
+                    className='mt-3 inline-flex w-full justify-center rounded-md border border-gray-300 bg-white px-4 py-2 text-base font-medium text-gray-700 shadow-sm hover:text-gray-500 focus:outline-none focus:ring-2 focus:ring-purple-500 focus:ring-offset-2 sm:mt-0 sm:w-auto sm:text-sm'
+                    onClick={() => setOpen(false)}
+                  >
+                    Close
+                  </button>
+                </div>
+              </DialogPanel>
+            </TransitionChild>
           </div>
-        )}
-        <div
-          className='flex px-4 py-5 hover:cursor-pointer hover:bg-gray-50 sm:px-6'
-          onClick={() => setShowResponse(!showResponse)}
-        >
-          {showResponse ? (
-            <ChevronDownIcon className='mr-2 h-5 w-5' aria-hidden='true' />
-          ) : (
-            <ChevronRightIcon className='mr-2 h-5 w-5' aria-hidden='true' />
-          )}{' '}
-          Response
         </div>
-        {showResponse && (
-          <div className='bg-gray-50 px-4 py-5 text-sm sm:p-6'>
-            <SyntaxHighlighter
-              language='javascript'
-              customStyle={{
-                backgroundColor: 'transparent',
-              }}
-            >
-              {JSON.stringify(
-                {
-                  id: 1,
-                  jsonrpc: '2.0',
-                  result:
-                    method.result.description == 'Null' ||
-                    !method.result.schema.examples
-                      ? []
-                      : [method.result.schema.examples[0]],
-                },
-                null,
-                2
-              )}
-            </SyntaxHighlighter>
-          </div>
-        )}
-      </div>
-    </div>
+      </Dialog>
+    </Transition>
   );
-};
-
-const getExampleRequest = (pkg: string, method: Method): string => {
-  return JSON.stringify(
-    {
-      id: 1,
-      jsonrpc: '2.0',
-      method: pkg + '.' + method.name,
-      params: method.params.map((param) =>
-        param.schema && param.schema.examples
-          ? param.schema.examples[0]
-          : undefined
-      ),
-    },
-    null,
-    2
-  );
-};
-
-const getExampleResponse = (method: Method): string => {
-  return JSON.stringify(
-    {
-      id: 1,
-      jsonrpc: '2.0',
-      result:
-        method.result.description == 'Null' || !method.result.schema.examples
-          ? []
-          : [method.result.schema.examples[0]],
-    },
-    null,
-    2
-  );
-};
-
-const sendRequest = async (
-  request: string,
-  hostname: string,
-  authtoken: string
-): Promise<object> => {
-  if (hostname == '') {
-    hostname = 'http://localhost:26658';
-  } else {
-    hostname = 'http://' + hostname;
-  }
-
-  if (authtoken != '') {
-    axios.defaults.headers.common['Authorization'] = `Bearer ${authtoken}`;
-  }
-
-  const { data } = await axios.post(hostname, JSON.parse(request));
-  return data;
 };
